+++ README.md
@@ -68,6 +68,7 @@ where each folder contains the dataset related to each task.
 - **[MinkowskiEngine](https://github.com/nicolas-chaulet/torch-points3d/tree/master/torch_points3d/modules/MinkowskiEngine)** from Christopher Choy _et al._: [4D Spatio-Temporal ConvNets: Minkowski Convolutional Neural Networks](https://arxiv.org/abs/1904.08755) (CVPR19)
 - **[VoteNet](https://github.com/nicolas-chaulet/torch-points3d/tree/master/torch_points3d/models/object_detection/votenet.py)** from Charles R. Qi _et al._: [Deep Hough Voting for 3D Object Detection in Point Clouds](https://arxiv.org/abs/1904.09664) (ICCV 19)
 - **[FCGF](https://github.com/chrischoy/FCGF)** from Christopher Choy _et al._: [Fully Convolutional Geometric Features](https://node1.chrischoy.org/data/publications/fcgf/fcgf.pdf) (ICCV'19)
+- **[PointGroup](https://github.com/Jia-Research-Lab/PointGroup)** from Li Jiang _et al._: [PointGroup: Dual-Set Point Grouping for 3D Instance Segmentation](https://arxiv.org/abs/2004.01658)
 
 ## Available datasets
 
@@ -352,9 +353,6 @@ Ensure that at least PyTorch 1.4.0 is installed and verify that `cuda/bin` and `
 ```
 $ python -c "import torch; print(torch.__version__)"
 >>> 1.4.0
-
-$ python -c "import torch; print(torch.__version__)"
->>> 1.1.0
 
 $ echo $PATH
 >>> /usr/local/cuda/bin:...
