+++ test/test_pointgroup.py
@@ -6,7 +6,7 @@ import torch
 ROOT = os.path.join(os.path.dirname(os.path.realpath(__file__)), "..")
 sys.path.insert(0, ROOT)
 
-from torch_points3d.models.panoptic.pointgroup import offset_loss, instance_iou_loss
+from torch_points3d.core.losses import offset_loss, instance_iou_loss
 
 
 class TestPointGroupLosses(unittest.TestCase):
