+++ torch_points3d/core/losses/__init__.py
@@ -2,6 +2,7 @@ import sys
 
 from .losses import *
 from .huber_loss import *
+from .panoptic_losses import *
 from pytorch_metric_learning.miners import *
 from pytorch_metric_learning.losses import *
